Add shared GroupTimeSeriesSplit utility and refactor scripts
